+++ JobIntroduction/src/main/java/com/oujava/service/impl/RatingServiceImpl.java
@@ -6,6 +6,7 @@ package com.oujava.service.impl;
 
 import com.oujava.pojo.Rating;
 import com.oujava.repository.RatingRepository;
+import com.oujava.repository.UserRepository;
 import com.oujava.service.RatingService;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,19 +17,40 @@ import org.springframework.stereotype.Service;
  * @author trann
  */
 @Service
-public class RatingServiceImpl implements RatingService{
+public class RatingServiceImpl implements RatingService {
 
     @Autowired
     private RatingRepository ratingRepo;
-    
+
+    @Autowired
+    private UserRepository userRepository;
+
     @Override
-    public List<Rating> getAllRating() {
-        return this.ratingRepo.getAllRating();
+    public List<Rating> getRatingOfCusId(int id) {
+        return ratingRepo.getRatingOfCusId(id);
+
     }
 
     @Override
-    public void deleteRatingById(int id) {
-        ratingRepo.deleteRatingById(id);
+    public Boolean addOrUpdateRating(Rating rating) {
+        if (!rating.getComment().isEmpty() || !rating.getComment().equals("")) {
+            if (!rating.getComment().isEmpty()) {
+                return ratingRepo.addOrUpdateRating(rating);
+            }else{
+                return false;
+            }
+        } else {
+            return false;
+        }
     }
-    
+
+    @Override
+    public Boolean deleteRatingById(int id) {
+        if(id >= 1){
+            return ratingRepo.deleteRatingById(id);
+        }else{
+            return false;
+        }
+    }
+
 }
